// types.ts
export interface Product {
  id: number;
  name: string;
  description: string;
  price: number;
  imageUrl: string;
  category: string;
}

// A product in the shopping cart
export interface CartItem extends Product {
  quantity: number;
}


// Type for the product generated by Gemini API
export interface GeneratedProduct {
  productName: string;
  description: string;
  priceInINR: number;
}